qr to the link
In the updated generateQR() function, we're generating a link that includes the description and image values as URL parameters. We're using the encodeURIComponent() function to properly encode the values, in case they contain special characters that could cause issues in a URL.

We're then passing the link to the QRCode constructor to generate the QR code.

When the QR code is scanned, the user will be taken to the URL with the description and image values as parameters. You can then retrieve the values from the URL on the server side to display the description and image.